requirements.txt
streamlit
python-dotenv
supabase
requests


